XCUtils extracted from XC files
